refact: split wallet codes
