+++ packages/wallet/src/index.ts
@@ -1,3 +1,2 @@
 export * from './wallets';
-export * from './wallet-locked';
 export * as TestUtils from './test-utils';
