+++ packages/wallet/src/wallets.ts
@@ -1,308 +1,21 @@
 //
-// Because wallet have a cycle dependency it's not possible
-// To split the two classes in different files
+// Because WalletLocked and WalletUnlocked has a cycle dependency
+// it's not possible to split the two classes in different files
 //
 /* eslint-disable max-classes-per-file */
 import type { BytesLike } from '@ethersproject/bytes';
-import { hexlify } from '@ethersproject/bytes';
-import type { InputValue } from '@fuel-ts/abi-coder';
-import { Address, addressify } from '@fuel-ts/address';
-import { NativeAssetId } from '@fuel-ts/constants';
-import { hashMessage, hashTransaction } from '@fuel-ts/hasher';
 import { HDWallet } from '@fuel-ts/hdwallet';
-import { AbstractWallet } from '@fuel-ts/interfaces';
-import type { AbstractAddress, AbstractPredicate } from '@fuel-ts/interfaces';
-import type { BigNumberish, BN } from '@fuel-ts/math';
 import { Mnemonic } from '@fuel-ts/mnemonic';
-import type {
-  TransactionResponse,
-  TransactionRequestLike,
-  CallResult,
-  TransactionRequest,
-  Coin,
-  CoinQuantityLike,
-  CoinQuantity,
-  BuildPredicateOptions,
-  TransactionResult,
-  Message,
-} from '@fuel-ts/providers';
-import { ScriptTransactionRequest, Provider, transactionRequestify } from '@fuel-ts/providers';
 import { Signer } from '@fuel-ts/signer';
-import { MAX_GAS_PER_TX } from '@fuel-ts/transactions';
 
-import { FUEL_NETWORK_URL } from './constants';
+import { BaseWalletLocked } from './base-locked-wallet';
+import { BaseWalletUnlocked } from './base-unlocked-wallet';
 import type { GenerateOptions } from './types/GenerateOptions';
 
 /**
  * WalletLocked
  */
-export class WalletLocked extends AbstractWallet {
-  private readonly _address: AbstractAddress;
-
-  provider: Provider;
-
-  constructor(publicKey: string | AbstractAddress, provider: string | Provider = FUEL_NETWORK_URL) {
-    super();
-    this.provider = this.connect(provider);
-    if (typeof publicKey === 'string') {
-      this._address = Address.fromString(publicKey);
-    } else {
-      this._address = addressify(publicKey);
-    }
-  }
-
-  get address(): AbstractAddress {
-    return this._address;
-  }
-
-  /**
-   * Change provider connection
-   */
-  connect(provider: string | Provider) {
-    if (!provider) {
-      throw new Error('Provider is required');
-    } else if (typeof provider === 'string') {
-      this.provider = new Provider(provider);
-    } else {
-      this.provider = provider;
-    }
-    return this.provider;
-  }
-
-  /**
-   * Returns coins satisfying the spend query.
-   */
-  async getCoinsToSpend(
-    quantities: CoinQuantityLike[],
-    /** IDs of coins to exclude */
-    excludedIds?: BytesLike[]
-  ): Promise<Coin[]> {
-    return this.provider.getCoinsToSpend(this.address, quantities, excludedIds);
-  }
-
-  /**
-   * Gets coins owned by the wallet address.
-   */
-  async getCoins(): Promise<Coin[]> {
-    const coins = [];
-
-    const pageSize = 9999;
-    let cursor;
-    // eslint-disable-next-line no-unreachable-loop
-    for (;;) {
-      const pageCoins = await this.provider.getCoins(this.address, undefined, {
-        first: pageSize,
-        after: cursor,
-      });
-
-      coins.push(...pageCoins);
-
-      const hasNextPage = pageCoins.length >= pageSize;
-      if (!hasNextPage) {
-        break;
-      }
-
-      // TODO: implement pagination
-      throw new Error(`Wallets with more than ${pageSize} coins are not yet supported`);
-    }
-
-    return coins;
-  }
-
-  /**
-   * Gets messages owned by the wallet address.
-   */
-  async getMessages(): Promise<Message[]> {
-    const messages = [];
-
-    const pageSize = 9999;
-    let cursor;
-    // eslint-disable-next-line no-unreachable-loop
-    for (;;) {
-      const pageMessages = await this.provider.getMessages(this.address, {
-        first: pageSize,
-        after: cursor,
-      });
-
-      messages.push(...pageMessages);
-
-      const hasNextPage = pageMessages.length >= pageSize;
-      if (!hasNextPage) {
-        break;
-      }
-
-      // TODO: implement pagination
-      throw new Error(`Wallets with more than ${pageSize} messages are not yet supported`);
-    }
-
-    return messages;
-  }
-
-  /**
-   * Gets balance for the given asset.
-   */
-  async getBalance(assetId: BytesLike = NativeAssetId): Promise<BN> {
-    const amount = await this.provider.getBalance(this.address, assetId);
-    return amount;
-  }
-
-  /**
-   * Gets balances.
-   */
-  async getBalances(): Promise<CoinQuantity[]> {
-    const balances = [];
-
-    const pageSize = 9999;
-    let cursor;
-    // eslint-disable-next-line no-unreachable-loop
-    for (;;) {
-      const pageBalances = await this.provider.getBalances(this.address, {
-        first: pageSize,
-        after: cursor,
-      });
-
-      balances.push(...pageBalances);
-
-      const hasNextPage = pageBalances.length >= pageSize;
-      if (!hasNextPage) {
-        break;
-      }
-
-      // TODO: implement pagination
-      throw new Error(`Wallets with more than ${pageSize} balances are not yet supported`);
-    }
-
-    return balances;
-  }
-
-  /**
-   * Adds coins to the transaction enough to fund it.
-   */
-  async fund<T extends TransactionRequest>(request: T): Promise<void> {
-    const fee = request.calculateFee();
-    const coins = await this.getCoinsToSpend([fee]);
-
-    request.addCoins(coins);
-  }
-
-  /**
-   * Returns coins satisfying the spend query.
-   */
-  async transfer(
-    /** Address of the destination */
-    destination: AbstractAddress,
-    /** Amount of coins */
-    amount: BigNumberish,
-    /** Asset ID of coins */
-    assetId: BytesLike = NativeAssetId,
-    /** Tx Params */
-    txParams: Pick<TransactionRequestLike, 'gasLimit' | 'gasPrice' | 'maturity'> = {}
-  ): Promise<TransactionResponse> {
-    const params = { gasLimit: MAX_GAS_PER_TX, ...txParams };
-    const request = new ScriptTransactionRequest(params);
-    request.addCoinOutput(destination, amount, assetId);
-    const fee = request.calculateFee();
-    let quantities: CoinQuantityLike[] = [];
-    if (fee.assetId === hexlify(assetId)) {
-      fee.amount.add(amount);
-      quantities = [fee];
-    } else {
-      quantities = [[amount, assetId], fee];
-    }
-    const coins = await this.getCoinsToSpend(quantities);
-    request.addCoins(coins);
-
-    return this.sendTransaction(request);
-  }
-
-  /**
-   * Populates witnesses signature and send it to the network using `provider.sendTransaction`.
-   *
-   * @param transactionRequest - TransactionRequest
-   * @returns TransactionResponse
-   */
-  async sendTransaction(
-    transactionRequestLike: TransactionRequestLike
-  ): Promise<TransactionResponse> {
-    const transactionRequest = transactionRequestify(transactionRequestLike);
-    return this.provider.sendTransaction(transactionRequest);
-  }
-
-  /**
-   * Populates witnesses signature and send a call it to the network using `provider.call`.
-   *
-   * @param transactionRequest - TransactionRequest
-   * @returns CallResult
-   */
-  async simulateTransaction(transactionRequestLike: TransactionRequestLike): Promise<CallResult> {
-    const transactionRequest = transactionRequestify(transactionRequestLike);
-    return this.provider.simulate(transactionRequest);
-  }
-
-  async buildPredicateTransaction(
-    predicateAddress: AbstractAddress,
-    amountToPredicate: BigNumberish,
-    assetId: BytesLike = NativeAssetId,
-    predicateOptions?: BuildPredicateOptions
-  ): Promise<ScriptTransactionRequest> {
-    const options = {
-      fundTransaction: true,
-      ...predicateOptions,
-    };
-    const request = new ScriptTransactionRequest({
-      gasLimit: MAX_GAS_PER_TX,
-      ...options,
-    });
-
-    // output is locked behind predicate
-    request.addCoinOutput(predicateAddress, amountToPredicate, assetId);
-
-    const requiredCoinQuantities: CoinQuantityLike[] = [];
-    if (options.fundTransaction) {
-      requiredCoinQuantities.push(request.calculateFee());
-    }
-
-    if (requiredCoinQuantities.length) {
-      const coins = await this.getCoinsToSpend(requiredCoinQuantities);
-      request.addCoins(coins);
-    }
-
-    return request;
-  }
-
-  async submitPredicate(
-    predicateAddress: AbstractAddress,
-    amountToPredicate: BigNumberish,
-    assetId: BytesLike = NativeAssetId,
-    options?: BuildPredicateOptions
-  ): Promise<TransactionResult<'success'>> {
-    const request = await this.buildPredicateTransaction(
-      predicateAddress,
-      amountToPredicate,
-      assetId,
-      options
-    );
-    const response = await this.sendTransaction(request);
-    return response.waitForResult();
-  }
-
-  async submitSpendPredicate(
-    predicate: AbstractPredicate,
-    amountToSpend: BigNumberish,
-    predicateData?: InputValue[],
-    assetId: BytesLike = NativeAssetId,
-    options?: BuildPredicateOptions
-  ): Promise<TransactionResult<'success'>> {
-    return this.provider.submitSpendPredicate(
-      predicate,
-      amountToSpend,
-      this.address,
-      predicateData,
-      assetId,
-      options
-    );
-  }
-
+export class WalletLocked extends BaseWalletLocked {
   unlock(privateKey: BytesLike) {
     // eslint-disable-next-line @typescript-eslint/no-use-before-define
     return new WalletUnlocked(privateKey, this.provider);
@@ -312,97 +25,10 @@ export class WalletLocked extends AbstractWallet {
 /**
  * WalletUnlocked
  */
-export class WalletUnlocked extends WalletLocked {
-  /* default HDWallet path */
-  static defaultPath = "m/44'/1179993420'/0'/0/0";
-
-  provider: Provider;
-
-  readonly signer: () => Signer;
-
-  constructor(privateKey: BytesLike, provider: string | Provider = FUEL_NETWORK_URL) {
-    const signer = new Signer(privateKey);
-    super(signer.address, provider);
-    this.signer = () => signer;
-    this.provider = this.connect(provider);
-  }
-
-  get privateKey(): string {
-    return this.signer().privateKey;
-  }
-
-  get publicKey(): string {
-    return this.signer().publicKey;
-  }
-
+export class WalletUnlocked extends BaseWalletUnlocked {
   lock(): WalletLocked {
+    this.signer = () => new Signer('0x00');
     return new WalletLocked(this.address, this.provider);
-  }
-
-  /**
-   * Sign message with wallet instance privateKey
-   *
-   * @param message - Message
-   * @returns Promise<string> - Signature a ECDSA 64 bytes
-   */
-  async signMessage(message: string): Promise<string> {
-    return this.signer().sign(hashMessage(message));
-  }
-
-  /**
-   * Sign transaction with wallet instance privateKey
-   *
-   * @param transactionRequestLike - TransactionRequestLike
-   * @returns string - Signature a ECDSA 64 bytes
-   */
-  async signTransaction(transactionRequestLike: TransactionRequestLike): Promise<string> {
-    const transactionRequest = transactionRequestify(transactionRequestLike);
-    const hashedTransaction = hashTransaction(transactionRequest);
-    const signature = this.signer().sign(hashedTransaction);
-
-    return signature;
-  }
-
-  async populateTransactionWitnessesSignature(transactionRequestLike: TransactionRequestLike) {
-    const transactionRequest = transactionRequestify(transactionRequestLike);
-    const signedTransaction = await this.signTransaction(transactionRequest);
-
-    transactionRequest.updateWitnessByOwner(this.address, signedTransaction);
-
-    return transactionRequest;
-  }
-
-  /**
-   * Populates witnesses signature and send it to the network using `provider.sendTransaction`.
-   *
-   * @param transactionRequest - TransactionRequest
-   * @returns TransactionResponse
-   */
-  async sendTransaction(
-    transactionRequestLike: TransactionRequestLike
-  ): Promise<TransactionResponse> {
-    const transactionRequest = transactionRequestify(transactionRequestLike);
-
-    return this.provider.sendTransaction(
-      await this.populateTransactionWitnessesSignature(transactionRequest)
-    );
-  }
-
-  /**
-   * Populates witnesses signature and send a call it to the network using `provider.call`.
-   *
-   * @param transactionRequest - TransactionRequest
-   * @returns CallResult
-   */
-  async simulateTransaction(transactionRequestLike: TransactionRequestLike): Promise<CallResult> {
-    const transactionRequest = transactionRequestify(transactionRequestLike);
-
-    return this.provider.call(
-      await this.populateTransactionWitnessesSignature(transactionRequest),
-      {
-        utxoValidation: true,
-      }
-    );
   }
 
   /**
